test : ClubController unit, integration test 파일 생성
